Added Modularity
separated page classes into separate files where i then imported into the hompage.dart file
